requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
pdfminer.six>=20240716
pdfplumber>=0.11.4
PyPDF2>=3.0.1
pytesseract>=0.3.13
python-docx>=1.1.2
pandas>=2.2.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
SQLAlchemy>=2.0.34
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
httpx>=0.27.2
pydantic>=2.9.2
PyYAML>=6.0.2
